+++ include/types/ScopedXLOPER12.h
@@ -1,10 +1,6 @@
 #pragma once
 
-#ifdef _WIN32
 #include <windows.h>
-#else
-#include "types/win_compat.h"
-#endif
 
 #include "types/xlcall.h"
 #include <vector>