+++ tests/test_fixes.cpp
@@ -7,9 +7,7 @@
 #include "types/utility.h"
 #include "types/mem.h"
 
-#ifdef _WIN32
 HINSTANCE g_hModule = NULL;
-#endif
 
 extern "C" void __stdcall xlAutoFree12(LPXLOPER12 p);