+++ tests/test_security_regressions.cpp
@@ -3,10 +3,8 @@
 #include <string>
 #include <limits>
 
-#ifdef _WIN32
 #include <windows.h>
 HINSTANCE g_hModule = NULL;
-#endif
 
 #include "types/converters.h"
 #include "types/mem.h"